requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.2.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
cryptography==35.0.0
cycler==0.11.0
Django==4.1.3
djangorestframework==3.14.0
fake-useragent==0.1.11
fonttools==4.28.1
gunicorn==20.1.0
h11==0.12.0
idna==3.4
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
outcome==1.1.0
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pyfiglet==0.8.post1
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
selenium==4.0.0b4
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
tzdata==2022.6
undetected-chromedriver==3.0.3
urllib3==1.26.12
websockets==10.0
whitenoise==6.2.0
wsproto==1.0.0
